:truck: [#1472] Move action files to action folder
